Split the graphQLStore in repositories
