+++ shared/domain/src/commonMain/kotlin/fr/androidmakers/domain/repo/PartnersRepository.kt
@@ -4,6 +4,5 @@ import fr.androidmakers.domain.model.Partner
 import kotlinx.coroutines.flow.Flow
 
 interface PartnersRepository {
-
   fun getPartners(): Flow<Result<List<Partner>>>
 }
